requirements.txt
camelot-py[cv]
Flask
pandas